任务管理重构
