+++ api/project/project.py
@@ -4,21 +4,18 @@ import asyncio
 from bson import ObjectId
 from fastapi import APIRouter, Depends, BackgroundTasks
 
-from api.task.util import delete_asset, create_scan_task
+from api.project.handler import update_project, delete_asset_project_handler
+from api.project.scan import scheduler_project
+from api.task.util import delete_asset
 from api.users import verify_token
-from motor.motor_asyncio import AsyncIOMotorCursor
-
-from core.config import Project_List
-from core.db import get_mongo_db
 from core.redis_handler import refresh_config, get_redis_pool
-from loguru import logger
 from core.util import *
 from core.apscheduler_handler import scheduler
 
 router = APIRouter()
 
 
-@router.post("/project/data")
+@router.post("/data")
 async def get_projects_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token),
                             background_tasks: BackgroundTasks = BackgroundTasks()):
     background_tasks.add_task(update_project_count)
@@ -83,7 +80,7 @@ async def get_projects_data(request_data: dict, db=Depends(get_mongo_db), _: dic
     }
 
 
-@router.get("/project/all")
+@router.get("/all")
 async def get_projects_all(db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     try:
         pipeline = [
@@ -136,7 +133,7 @@ async def update_project_count():
         await asyncio.gather(*fetch_tasks)
 
 
-@router.post("/project/content")
+@router.post("/content")
 async def get_project_content(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     project_id = request_data.get("id")
     if not project_id:
@@ -172,7 +169,7 @@ async def get_project_content(request_data: dict, db=Depends(get_mongo_db), _: d
     return {"code": 200, "data": result}
 
 
-@router.post("/project/add")
+@router.post("/add")
 async def add_project_rule(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token),
                            background_tasks: BackgroundTasks = BackgroundTasks()):
     try:
@@ -232,7 +229,7 @@ async def add_project_rule(request_data: dict, db=Depends(get_mongo_db), _: dict
         return {"message": "error", "code": 500}
 
 
-@router.post("/project/delete")
+@router.post("/delete")
 async def delete_project_rules(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token),
                                background_tasks: BackgroundTasks = BackgroundTasks()):
     try:
@@ -265,7 +262,7 @@ async def delete_project_rules(request_data: dict, db=Depends(get_mongo_db), _: 
         return {"message": "error", "code": 500}
 
 
-@router.post("/project/update")
+@router.post("/update")
 async def update_project_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token),
                               background_tasks: BackgroundTasks = BackgroundTasks()):
     try:
@@ -356,125 +353,3 @@ async def update_project_data(request_data: dict, db=Depends(get_mongo_db), _: d
         # Handle exceptions as needed
         return {"message": "error", "code": 500}
 
-async def update_project(root_domain, project_id, change=False):
-    asset_collection_list = {
-                        'asset': ["url", "host", "ip"],
-                        'subdomain': ["host", "ip"],
-                        'DirScanResult': ["url"],
-                        'vulnerability': ["url"],
-                        'SubdoaminTakerResult': ["input"],
-                        'PageMonitoring': ["url"],
-                        'SensitiveResult': ["url"],
-                        'UrlScan': ["input"],
-                        'crawler': ["url"]}
-    async for db in get_mongo_db():
-        for a in asset_collection_list:
-            if change:
-                await asset_update_project(root_domain, asset_collection_list[a], a, db, project_id)
-            else:
-                await asset_add_project(root_domain, asset_collection_list[a], a, db, project_id)
-
-
-async def asset_add_project(root_domain, db_key, doc_name, db, project_id):
-    regex_patterns = [f".*{domain}.*" for domain in root_domain]
-    pattern = "|".join(regex_patterns)
-    # 构建查询条件
-    query = {
-        "$and": [
-            {
-                "$or": [
-                    {key: {"$regex": pattern, "$options": "i"}} for key in db_key
-                ]
-            },
-            {"project": {"$exists": True, "$eq": ""}}
-        ]
-    }
-    update_query = {
-        "$set": {
-            "project": project_id
-        }
-    }
-    result = await db[doc_name].update_many(query, update_query)
-    # 打印更新的文档数量
-    logger.info(f"Updated {doc_name} {result.modified_count} documents")
-
-
-async def asset_update_project(root_domain, db_key, doc_name, db, project_id):
-    regex_patterns = [f".*{domain}.*" for domain in root_domain]
-    pattern = "|".join(regex_patterns)
-    # 构建查询条件
-    query = {
-        "$and": [
-            {"project": project_id},
-            {
-                "$nor": [
-                    {key: {"$regex": pattern, "$options": "i"}} for key in db_key
-                ]
-            }
-        ]
-    }
-    update_query = {
-        "$set": {
-            "project": ""
-        }
-    }
-    result = await db[doc_name].update_many(query, update_query)
-    # 打印更新的文档数量
-    logger.info(f"Updated {doc_name} {result.modified_count} documents to null ")
-    await asset_add_project(root_domain, db_key, doc_name, db, project_id)
-
-
-async def delete_asset_project(db, collection, project_id):
-    try:
-        query = {"project": project_id}
-
-        cursor = db[collection].find(query)
-
-        async for document in cursor:
-            await db[collection].update_one({"_id": document["_id"]}, {"$set": {"project": ""}})
-
-    except Exception as e:
-        logger.error(f"delete_asset_project error:{e}")
-
-
-async def delete_asset_project_handler(project_id):
-    async for db in get_mongo_db():
-        asset_collection_list = ['asset', 'subdomain', 'DirScanResult', 'vulnerability', 'SubdoaminTakerResult',
-                                 'PageMonitoring', 'SensitiveResult', 'UrlScan', 'crawler']
-        for c in asset_collection_list:
-            await delete_asset_project(db, c, project_id)
-
-
-async def scheduler_project(id):
-    logger.info(f"Scheduler project {id}")
-    async for db in get_mongo_db():
-        async for redis in get_redis_pool():
-            job = scheduler.get_job(id)
-            task_id = generate_random_string(15)
-            if job:
-                next_time = job.next_run_time
-                formatted_time = next_time.strftime("%Y-%m-%d %H:%M:%S")
-                doc = await db.ScheduledTasks.find_one({"id": id})
-                run_id_last = doc.get("runner_id", "")
-                if run_id_last != "":
-                    progresskeys = await redis.keys(f"TaskInfo:progress:{run_id_last}:*")
-                    for pgk in progresskeys:
-                        await redis.delete(pgk)
-                update_document = {
-                    "$set": {
-                        "lastTime": get_now_time(),
-                        "nextTime": formatted_time,
-                        "runner_id": task_id
-                    }
-                }
-                await db.ScheduledTasks.update_one({"id": id}, update_document)
-            query = {"_id": ObjectId(id)}
-            doc = await db.project.find_one(query)
-            targetList = []
-            target_data = await db.ProjectTargetData.find_one({"id": id})
-            for t in target_data.get('target', '').split("\n"):
-                t.replace("http://", "").replace("https://", "")
-                t = t.strip("\n").strip("\r").strip()
-                if t != "" and t not in targetList:
-                    targetList.append(t)
-            await create_scan_task(doc, task_id, targetList, redis)
