+++ api/task/scheduled.py
@@ -4,12 +4,15 @@
 # @contact   : rainy-autumn@outlook.com
 # @time      : 2024/4/28 20:58
 # -------------------------------------------
+import traceback
+
 from apscheduler.events import JobSubmissionEvent, EVENT_JOB_MAX_INSTANCES, EVENT_JOB_SUBMITTED
 from apscheduler.executors.base import MaxInstancesReachedError
 from bson import ObjectId
 from fastapi import APIRouter, Depends
 from pytz import utc
 
+from api.task.handler import scheduler_scan_task
 from api.users import verify_token
 from motor.motor_asyncio import AsyncIOMotorCursor
 
@@ -184,7 +187,7 @@ async def create_page_monitoring_task():
                 await redis.rpush(f"NodeTask:{name}", json.dumps(add_redis_task_data))
 
 
-@router.post("/scheduled/task/pagemonit/data")
+@router.post("/pagemonit/data")
 async def get_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     try:
         search_query = request_data.get("search", "")
@@ -220,7 +223,7 @@ async def get_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_m
         return {"message": "error", "code": 500}
 
 
-@router.post("/scheduled/task/pagemonit/update")
+@router.post("/pagemonit/update")
 async def update_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     try:
         if not request_data:
@@ -257,7 +260,7 @@ async def update_scheduled_task_pagemonit_data(request_data: dict, db=Depends(ge
         return {"message": "error", "code": 500}
 
 
-@router.post("/scheduled/task/pagemonit/delete")
+@router.post("/pagemonit/delete")
 async def delete_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     try:
         # Extract the list of IDs from the request_data dictionary
@@ -281,7 +284,7 @@ async def delete_scheduled_task_pagemonit_data(request_data: dict, db=Depends(ge
         return {"message": "error", "code": 500}
 
 
-@router.post("/scheduled/task/pagemonit/add")
+@router.post("/pagemonit/add")
 async def add_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
     try:
         if not request_data:
@@ -305,4 +308,54 @@ async def add_scheduled_task_pagemonit_data(request_data: dict, db=Depends(get_m
     except Exception as e:
         logger.error(str(e))
         # Handle exceptions as needed
+        return {"message": "error", "code": 500}
+
+
+@router.post("/update")
+async def update_task_data(request_data: dict, db=Depends(get_mongo_db), _: dict = Depends(verify_token)):
+    try:
+        # Get the ID from the request data
+        task_id = request_data.get("id")
+        hour = request_data.get("hour")
+        # Check if ID is provided
+        if not task_id:
+            return {"message": "ID is missing in the request data", "code": 400}
+        query = {"id": ObjectId(task_id)}
+        doc = await db.ScheduledTasks.find_one(query)
+        oldScheduledTasks = doc["state"]
+        old_hour = doc["hour"]
+        newScheduledTasks = request_data.get("scheduledTasks")
+        if oldScheduledTasks is False:
+            # 原本是关闭的
+            # 如果重新开启，则开启计划任务
+            if newScheduledTasks:
+                scheduler.add_job(scheduler_scan_task, 'interval', hours=hour, args=[task_id], id=str(task_id),
+                                  jobstore='mongo')
+                await db.ScheduledTasks.update_one({"id": task_id}, {"$set": {'state': True}})
+            # 如果新的还是关闭，则不用管
+        else:
+            # 原本是开启的
+            # 如果重新提交的是开启的
+            if newScheduledTasks:
+                # 如果旧的时间间隔和新的时间剑客不同删除旧的然后按照新的时间重新提交，否则无需改变
+                if old_hour != hour:
+                    scheduler.remove_job(task_id)
+                    scheduler.add_job(scheduler_scan_task, 'interval', hours=hour, args=[task_id], id=str(task_id),
+                                      jobstore='mongo')
+            else:
+                # 重新提交的是关闭的，则移除计划任务
+                scheduler.remove_job(task_id)
+        request_data.pop("id")
+        update_document = {
+            "$set": request_data
+        }
+        result = await db.ScheduledTasks.update_one({"id": task_id}, update_document)
+        if result:
+            return {"message": "Task updated successfully", "code": 200}
+        else:
+            return {"message": "Failed to update data", "code": 404}
+
+    except Exception as e:
+        logger.error(str(e))
+        logger.error(traceback.format_exc())
         return {"message": "error", "code": 500}
