+++ main.py
@@ -1,14 +1,11 @@
-import subprocess
 import time
 
 from loguru import logger
 import uvicorn
-from motor.motor_asyncio import AsyncIOMotorGridFSBucket
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.staticfiles import StaticFiles
 
 from core.config import *
-from core.default import get_dirDict, get_domainDict, get_sensitive, PLUGINS
 from core.update import update14, update15
 
 set_config()
@@ -17,7 +14,6 @@ from core.db import get_mongo_db
 
 from starlette.requests import Request
 import asyncio
-from urllib.parse import urlparse
 from fastapi import FastAPI
 from fastapi.responses import FileResponse
 from fastapi.responses import JSONResponse
@@ -68,7 +64,7 @@ async def startup_db_client():
         if j.id == 'page_monitoring':
             find_page_m = True
     if not find_page_m:
-        from api.scheduled_tasks import get_page_monitoring_time, create_page_monitoring_task
+        from api.task.scheduled_tasks import get_page_monitoring_time, create_page_monitoring_task
         pat, flag = await get_page_monitoring_time()
         if flag:
             scheduler.add_job(create_page_monitoring_task, 'interval', hours=pat, id='page_monitoring',
@@ -85,12 +81,14 @@ async def http_exception_handler(request, exc):
 
 os.chdir(os.path.dirname(os.path.abspath(__file__)))
 
-from api import users, poc, configuration, fingerprint, node, project, task, \
-    page_monitoring, vulnerability, SubdoaminTaker, scheduled_tasks, notification, system, export, project_aggregation
+from api import users, poc, configuration, fingerprint, node, task, \
+    page_monitoring, vulnerability, SubdoaminTaker, notification, system, export, project_aggregation
+from api.task import scheduled_tasks
 
 from api.dictionary import router as dictionary_router
 from api.asset import router as asset_route
 from api.plugins import router as plugin_route
+from api.project import router as project_route
 
 app.include_router(plugin_route, prefix='/api')
 app.include_router(users.router, prefix='/api')
@@ -99,7 +97,7 @@ app.include_router(poc.router, prefix='/api')
 app.include_router(configuration.router, prefix='/api/configuration')
 app.include_router(fingerprint.router, prefix='/api')
 app.include_router(node.router, prefix='/api')
-app.include_router(project.router, prefix='/api')
+app.include_router(project_route, prefix='/api')
 app.include_router(task.router, prefix='/api')
 app.include_router(asset_route, prefix='/api')
 app.include_router(page_monitoring.router, prefix='/api')
